% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_longitudinal.R
\name{cv.fusedLongitudinal}
\alias{cv.fusedLongitudinal}
\title{cross validation for interaction detection for longitudinal outcomes using fused lasso}
\usage{
cv.fusedLongitudinal(x, y, family = c("gaussian", "binomial", "coxph"),
  lambda = numeric(0), nlambda = 100L, nfolds = 10L, foldid = NULL, ...)
}
\arguments{
\item{x}{input matrix.
Each row is an observation, each column corresponds to a covariate}

\item{y}{numeric response vector of length nobs.}

\item{family}{"gaussian" for least squares problems, "binomial" for binary response.
"coxph" for time-to-event outcomes}

\item{lambda}{tuning parameter values for lasso penalty}

\item{nlambda}{number of tuning parameter values - default is 100.}

\item{...}{other parameters to be passed to fusedLongitudinal}

\item{gamma}{ratio of fused lasso to lasso tuning parameter}
}
\value{
An object with S3 class "cv.fusedLongitudinal"
}
\description{
cross validation for interaction detection for longitudinal outcomes using fused lasso
}
\examples{
set.seed(123)
n.obs <- 100
n.vars <- 10

true.beta <- c(runif(3, -0.25, 0.25), rep(0, n.vars - 3))
true.beta.int <- c(rep(0, n.vars - 3), runif(3, -0.5, 0.5))

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
trt <- rbinom(n.obs, 1, 0.5)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta + trt * (x \%*\% true.beta.int)

fit <- cv.fusedLongitudinal(x = x, y = y)

}
