% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_response.R
\name{multioutcome}
\alias{multioutcome}
\title{Fit multivariate algorithm on the problem of multiple outcomes personalized medicine}
\usage{
multioutcome(Y, W, X, Trt, P, maxiter, mu, lossf = c("logistic", "square"),
  blearner = c("spline", "linear", "stump"))
}
\arguments{
\item{Y}{Outcome Matrix.
Each row is an observation, each column corresponds to a different outcome}

\item{W}{Weight matrix.
Each row is an observation, each column corresponds to a ooutcome.}

\item{X}{Covariate Matrix
Each row is an observation, each column corresponds to a covariate}

\item{maxiter}{Maximum number of iterations}

\item{mu}{stepsize in boosting}

\item{lossf}{Loss function to minimize. "logistic" for logistic loss, only works if all the outcomes are binary with 0 and 1. "square"
for square loss, it works for all type of responses.}

\item{blearner}{The base learner to use in boosting. "spline" for P-spline with B-spline basis. "linear" for linear model. "stump" for stump.}

\item{gamma}{ratio of fused lasso to lasso tuning parameter}
}
\value{
A object contraining information of the boosting algorithm
}
\description{
Fit multivariate algorithm on the problem of multiple outcomes personalized medicine
}
