% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_response.R
\name{predict.cvmultioutcome}
\alias{predict.cvmultioutcome}
\title{Predict a cvmultivariate object at the optimal number of iterations}
\usage{
\method{predict}{cvmultioutcome}(mb_mod, newdata, ...)
}
\arguments{
\item{mb_mod}{A object of class "cvmultivariate"}

\item{newdata}{New dataset with which to predict.}
}
\value{
Prediction of a multivariate object at optimal number of iterations
}
\description{
Predict a cvmultivariate object at the optimal number of iterations
}
