% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_longitudinal.R
\name{predict.fusedLongitudinal}
\alias{predict.fusedLongitudinal}
\title{Prediction method for fusedLongitudinal fitted objects}
\usage{
\method{predict}{fusedLongitudinal}(object, newx, s = NULL, type = c("link",
  "response", "coefficients", "nonzero", "class"), ...)
}
\arguments{
\item{object}{fitted "fusedLongitudinal" model object}

\item{newx}{Matrix of new values for x at which predictions are to be made. Must be a matrix; can be sparse as in Matrix package.
This argument is not used for type=c("coefficients","nonzero")}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required. Default is the entire sequence used to create
the model.}

\item{type}{Type of prediction required. Type == "link" gives the linear predictors for the "binomial" model; for "gaussian" models it gives the fitted values.
Type == "response" gives the fitted probabilities for "binomial". Type "coefficients" computes the coefficients at the requested values for s.
Type "class" applies only to "binomial" and produces the class label corresponding to the maximum probability.}

\item{...}{not used}
}
\value{
An object depending on the type argument
}
\description{
Prediction method for fusedLongitudinal fitted objects
}
\examples{
set.seed(123)
n.obs <- 100
n.vars <- 10

true.beta <- c(runif(3, -0.25, 0.25), rep(0, n.vars - 3))
true.beta.int <- c(rep(0, n.vars - 3), runif(3, -0.5, 0.5))

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
trt <- rbinom(n.obs, 1, 0.5)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta + trt * (x \%*\% true.beta.int)

fit <- fusedLongitudinal(x = x, y = y)

preds <- predict(fit, newx = x)


}
