% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_response.R
\name{predict.multioutcome}
\alias{predict.multioutcome}
\title{Predict a multivariate object}
\usage{
\method{predict}{multioutcome}(mb_mod, iter, newdata, ...)
}
\arguments{
\item{mb_mod}{A object of class "multivariate"}

\item{iter}{The number of iterations for prediction}

\item{newdata}{New dataset with which to predict.}
}
\value{
Prediction of a multivariate object at some iteration
}
\description{
Predict a multivariate object
}
