% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_longitudinal.R
\name{summary.fusedLongitudinal}
\alias{summary.fusedLongitudinal}
\title{Summary method for fusedLongitudinal fitted objects}
\usage{
summary.fusedLongitudinal(object, ...)
}
\arguments{
\item{object}{fitted "fusedLongitudinal" model object}

\item{...}{not used}
}
\value{
An object depending on the type argument
}
\description{
Summary method for fusedLongitudinal fitted objects
}
\examples{
set.seed(123)
n.obs <- 100
n.vars <- 10

true.beta <- c(runif(3, -0.25, 0.25), rep(0, n.vars - 3))
true.beta.int <- c(rep(0, n.vars - 3), runif(3, -0.5, 0.5))

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
trt <- rbinom(n.obs, 1, 0.5)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta + trt * (x \%*\% true.beta.int)

fit <- fusedLongitudinal(x = x, y = y)

summary(fit)


}
