% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_longitudinal.R
\name{summary.fusedLongitudinal}
\alias{summary.fusedLongitudinal}
\title{Summary method for fusedLongitudinal fitted objects}
\usage{
\method{summary}{fusedLongitudinal}(object, ...)
}
\arguments{
\item{object}{fitted "fusedLongitudinal" model object}

\item{...}{not used}
}
\value{
An object depending on the type argument
}
\description{
Summary method for fusedLongitudinal fitted objects
}
\examples{
set.seed(123)
nobs       <- 100
nobs.test  <- 1e5
nvars      <- 20
periods    <- 6
sd         <- 2

beta.nz <- rbind(c( 1,    1,    1,    1.5,  1.5,  1.5),
                 c(-1,   -1,   -1,   -0.5, -0.5, -0.5),
                 c( 1,    1,    1,   -1,   -1,   -1),
                 c( 1,    1,    1,    1,    1,    1),
                 c(-0.5, -0.5, -0.5, -0.5, -0.5, -0.5))

beta <- data.matrix(rbind(beta.nz, matrix(0, nvars - 5, periods)))

trt <- rbinom(nobs, 1, 0.5)
x   <- matrix(rnorm(nobs * nvars), ncol = nvars); colnames(x) <- paste0("V", 1:ncol(x))
y   <- x \%*\% (beta * 0.5) + (2 * trt - 1) * (x \%*\% beta) +
    matrix(rnorm(nobs * periods, sd = sd), ncol = periods)
y   <- apply(y, 2, function(yy) yy - (2 * trt - 1))

plot(x = NULL, xlim = c(1,6), ylim = range(y))
for (i in 1:nobs)
{
    lines(x = 1:6, y = y[i,], col = colors()[i+1])
}

x.list <- rep(list(x), periods)
y.list <- lapply(apply(y, 2, function(x) list(x) ), function(x) x[[1]])

fit <- fusedLongitudinal(x = x.list, y = y.list, trt = trt, gamma = c(0.05, 0.1, 1, 2, 5, 10))

summary(fit)


}
